021_sd_template create discrete TUs for audio and gui thread definitions
